(GH-14) Ensure Attr calls aren't made when console is redirected
Previously we were getting bad exceptions due to calls to Win32::Console::Attr
when the console was redirected to something other then the console, for
example:

    ruby myapp.rb | more

This patch moves the redirect? method into Win32::Console (previously this main
logic was in Win32::Console::ANSI) so that our initialization can use it and
avoid using dangerous calls in this state. I anticipate there are more cases,
not just Attr so this method call may come in handy.

Also - in the future redirect? may be useful for avoiding lots of unnecessary
processing as no color modifications should be made in this state at all.

I've also added the first few RSpec tests to stop this bug from coming back.
The tests are primarily focused on this problem alone, but this will give us a
baseline to grow our unit tests from.